+++ CHANGELOG.md
@@ -5,4 +5,8 @@
 - **RRV Ingestion**
   - Improved performance
   - Centered the recipe viewer in the middle of the screen
-  - Fixed pressing R and U over an item in the inventory not showing the recipe or usage.
+  - Fixed pressing R and U over an item in the inventory not showing the recipe or usage
+- **Chat Enhancements**
+  - Pressing the up arrow key no longer selects the chat tab; it now selects the last used command
+  - Fixed compact-only consecutive messages not working
+  - Added option to only compact messages sent in the last x minutes (default 10 min)